feat: Improve scripts and plugins
This commit includes several improvements to scripts and plugins.
In `wezterm/status.lua`, a check is added to ensure the string `s` is not
null before performing operations on it. This prevents potential runtime
errors.

A new script `nv-r` is added. This script checks if the `NVIM` environment
variable is set and launches Neovim accordingly. This provides flexibility
in launching Neovim in different environments.

In `traildots.plugin.zsh`, a local variable `i` is declared. This change
ensures that the variable `i` is not globally accessible, preventing
potential conflicts with other scripts.