feat(wezterm): improve status bar and tab title
This commit improves the status bar and tab title in wezterm. It adds
memory and CPU usage to the status bar. It also modifies the tab title
to include the number of panes, the title, and the current working
directory. The commit also cleans up the .zshrc and .zprofile files by
removing unnecessary lines and consolidating the plugin list.