feat(wezterm): update status bar and tab title symbols
This commit updates the symbols used in the status bar and tab title
in the Wezterm configuration. The powerline < symbol and the filled in
variant of the >| symbol are replaced with nerdfonts separators. The
memory and CPU usage indicators are also updated to use a different
separator. Additionally, the update window is no longer shown on
Windows, and the close tab button is hidden in tabs.